% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_buildurl}
\alias{api_buildurl}
\title{Build base URL}
\usage{
api_buildurl(endpoint, query = NA, database = NA, extension = "json")
}
\arguments{
\item{endpoint}{The endpoint, e.g. articles/import}

\item{query}{Query parameters for the endpoint}

\item{database}{The database name}

\item{extension}{Extension added to the URL path, defaults to json.}
}
\description{
Build base URL
}
