% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_job_create}
\alias{api_job_create}
\title{Create and execute a job}
\usage{
api_job_create(endpoint, params, database, payload = NULL)
}
\arguments{
\item{endpoint}{The endpoint supporting job creation}

\item{params}{Query parameters}

\item{database}{The selected database}

\item{payload}{The data posted to the job endpoint}
}
\value{
void
}
\description{
Create and execute a job
}
