% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch}
\alias{api_patch}
\title{Patch data}
\usage{
api_patch(data, database)
}
\arguments{
\item{data}{A dataframe with the column id (must be a a valid IRI path).
Additional columns such as norm_data will be written to the record.}

\item{database}{The database name}
}
\description{
Update records in the database using the API.
Existing records will be updated, missing records will be created.
The function supports uploading all data related to articles:
articles, sections, items, links, footnotes, properties, projects, users, types.
The IRI path in the ID column of the dataframe must contain the specific table name.
}
