% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch_articles}
\alias{api_patch_articles}
\title{Patch articles}
\usage{
api_patch_articles(articles, database)
}
\arguments{
\item{database}{The database name}

\item{sections}{A dataframe with the column id (must be a a valid IRI path).
Additional columns such as norm_data will be written to the article}
}
\description{
Update articles in the database using the API.
Existing articles will be updated, missing articles will be created.
}
