% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch_items}
\alias{api_patch_items}
\title{Patch items and create related properties, sections, articles and projects}
\usage{
api_patch_items(data, database)
}
\arguments{
\item{data}{A dataframe with the column id containing a valid item IRI path.
Additional columns such as norm_data will be written to the record.\preformatted{        Column names prefixed with "properties", "sections", "articles"
        and "projects" followed by a dot (e.g. "properties.id",
        "properties.lemma") indicate which other records will be patched.
}}

\item{database}{The database name}
}
\description{
Patch items and create related properties, sections, articles and projects
}
