% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch_properties}
\alias{api_patch_properties}
\title{Patch properties}
\usage{
api_patch_properties(database, propertytype, lemmata, irifragments = NA)
}
\arguments{
\item{database}{The database name}

\item{propertytype}{The property type (character)}

\item{lemmata}{A vector of lemmata (character).}

\item{irifragments}{Optional. A vector of IRI fragments (character).}
}
\description{
Update properties in the database using the API.
Existing properties will be updated, missing properties will be created.
}
