% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch_sections}
\alias{api_patch_sections}
\title{Patch sections}
\usage{
api_patch_sections(sections, database)
}
\arguments{
\item{sections}{A dataframe with the columns
article (must be a a valid IRI path) and
section (either a valid IRI path or a section type)
Additional columns such as name will be written to the section.}

\item{database}{The database name}
}
\description{
Update sections in the database using the API.
Existing sections will be updated, missing sections will be created.
Section IRI fragments will be derived from the article IRI fragments.
}
