% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch_types}
\alias{api_patch_types}
\title{Patch types}
\usage{
api_patch_types(types, database)
}
\arguments{
\item{types}{A dataframe with the column id (must be a a valid IRI path).
Additional columns such as norm_data will be written to the article}

\item{database}{The database name}
}
\description{
Update the types config in the database using the API.
Existing types will be updated, missing types will be created.
}
