% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch_wide}
\alias{api_patch_wide}
\title{Patch data and create related properties, items, sections, articles and projects}
\usage{
api_patch_wide(data, database)
}
\arguments{
\item{data}{A dataframe with the column id containing a valid IRI path.
Additional columns such as norm_data will be written to the record.\preformatted{        Column names prefixed with "properties", "items", "sections", "articles"
        and "projects" followed by a dot (e.g. "properties.id",
        "properties.lemma") will be extracted and patched as additional records
}}

\item{database}{The database name}
}
\description{
Patch data and create related properties, items, sections, articles and projects
}
