% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_silent}
\alias{api_silent}
\title{Set silent mode}
\usage{
api_silent(silent = F)
}
\arguments{
\item{silent}{Boolen}
}
\description{
In silent mode, all user prompts are automatically confirmed.
Be careful, this will skip the prompt to confirm operations
on the live server.
}
