% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_table}
\alias{api_table}
\title{Download tabular data}
\usage{
api_table(table, params = c(), db, maxpages = 1)
}
\arguments{
\item{table}{The table name (e.g. "items")}

\item{params}{A named list of query params}

\item{db}{The database name}

\item{maxpages}{Maximum number of pages to request.
Set to 1 for non-paginated tables.}
}
\description{
Download tabular data
}
