% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{api_transfer}
\alias{api_transfer}
\title{Transfer datasets between different databases}
\usage{
api_transfer(table, db_source, db_target, db_params = list())
}
\arguments{
\item{table}{Table name, e.g. "types"}

\item{db_source}{Source database name}

\item{db_target}{Target database name}

\item{db_params}{A parameter list for selecting the appropriate rows.
For example: params <- list("scopes" = "properties")}
}
\description{
Transfer datasets between different databases
}
