% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_rows_char}
\alias{bind_rows_char}
\title{Bind rows of dataframes even if column types differ}
\usage{
bind_rows_char(dataframes)
}
\arguments{
\item{dataframes}{A list of dataframes}

\item{A}{list of dataframes with common column types}
}
\description{
Convert columns that differ in type to character,
then bind rows.
}
