% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_db}
\alias{check_is_db}
\title{Check whether a value is a valid database name, and stop if not}
\usage{
check_is_db(value, msg = NULL)
}
\arguments{
\item{value}{A character value}

\item{msg}{A custom error message if the check fails}
}
\value{
boolean Whether the value is a valid database name
}
\description{
Check whether a value is a valid database name, and stop if not
}
\keyword{internal}
