% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_is_id}
\alias{check_is_id}
\title{Check whether a value is a valid table prefixed ID, and stop if not}
\usage{
check_is_id(value, msg = NULL)
}
\arguments{
\item{value}{A character value}

\item{msg}{A custom error message if the check fails}
}
\value{
boolean Whether the value is a valid table prefixed ID
}
\description{
Check whether a value is a valid table prefixed ID, and stop if not
}
\keyword{internal}
