% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft_compile}
\alias{craft_compile}
\title{Compile a crafted Epigraf table ready to patch into the database}
\usage{
craft_compile(ds)
}
\description{
Compile a crafted Epigraf table ready to patch into the database
}
