% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft_type_fields}
\alias{craft_type_fields}
\title{Craft field configuration}
\usage{
craft_type_fields(
  ds,
  col_name = field_name,
  cols_fields,
  col_id,
  col_mode = mode,
  type
)
}
\description{
Craft field configuration
}
