% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_annotations}
\alias{db_annotations}
\title{Get annotations from from items and links}
\usage{
db_annotations(db)
}
\arguments{
\item{db}{A connection object (object) or the database name (character)}
}
\value{
A dataframe containing links and items.
}
\description{
Get annotations from from items and links
}
\details{
Queries the database to join the items, properties, types, articles and links tables.
}
