% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_connect}
\alias{db_connect}
\title{Get a connection to a database}
\usage{
db_connect(db = NULL)
}
\arguments{
\item{db}{Name of the database as string.
Leave empty to use the database
name from the environment settings.}
}
\description{
Before you can use this function, call db_setup once
to set the connection parameters.
All parameters are stored in the environment.
}
