% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_databases}
\alias{db_databases}
\title{Get list of all databases}
\usage{
db_databases(epi = FALSE)
}
\arguments{
\item{epi}{Only keep databases with the epi-prefix.}
}
\description{
Get list of all databases
}
