% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_geolocations}
\alias{db_geolocations}
\title{Extract latitude and longitude values from items}
\usage{
db_geolocations(db, itemtype = "geolocations")
}
\arguments{
\item{itemtype}{The itemtype}
}
\value{
A data frame containing the geolocations
}
\description{
Extract latitude and longitude values from items
}
\details{
Geolocation data is extracted from the
JSON data contained in the value field.
}
