% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_get_codes}
\alias{db_get_codes}
\title{Get codes from database}
\usage{
db_get_codes(db)
}
\arguments{
\item{db}{A connection object (object) or the database name (character)}
}
\value{
A data frame containing the properties.
}
\description{
Get codes from database
}
\details{
Queries the database to get the properties.
The property types are added as root nodes
and if the lemma is empty, it is replaced by the name
}
