% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_get_codings}
\alias{db_get_codings}
\title{Get codings from database}
\usage{
db_get_codings(db)
}
\arguments{
\item{db}{A connection object (object) or the database name (character)}
}
\value{
A dataframe containing links and items.
}
\description{
Get codings from database
}
\details{
Queries the database to join the items, properties, types, articles and links tables.
}
