% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_name}
\alias{db_name}
\title{Get the database name from a connection object}
\usage{
db_name(con)
}
\arguments{
\item{con}{A connection object}
}
\description{
Get the database name from a connection object
}
