% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_setup}
\alias{db_setup}
\title{Save database connection settings to environment variables.
Environment variables are prefixed with "epi_" and used in db_connect()
to establish the connection.}
\usage{
db_setup(
  host = "localhost",
  port = 3306,
  username = "root",
  password = "root",
  database = ""
)
}
\description{
Save database connection settings to environment variables.
Environment variables are prefixed with "epi_" and used in db_connect()
to establish the connection.
}
