% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decode}
\alias{decode}
\title{Converts "b" using the "base"}
\usage{
decode(b, base)
}
\arguments{
\item{b}{The number to convert}

\item{base}{The base}
}
\description{
APL-decode / APL-base "\emph{|}", pw10/02
}
