% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{df_to_ram}
\alias{df_to_ram}
\title{Map a data frame to the Relational Article Model}
\usage{
df_to_ram(
  df,
  item = c(),
  section = c(),
  article = c(),
  project = c(),
  vals = list(),
  compile = FALSE
)
}
\arguments{
\item{df}{A data frame with the source data.}

\item{item}{A named list. Names are RAM columns, values source columns.}

\item{section}{A named list. Names are RAM columns, values source columns.}

\item{article}{A named list. Names are RAM columns, values source columns.}

\item{project}{A named list. Names are RAM columns, values source columns.}

\item{vals}{A named list with fixed values for the rows.
Names on the first level correspond to the table (item, section, article, project).
Names on the second level correspond the the field names.}

\item{compile}{Whether to return compiled rows or store them in the attributes of the input data frame.}
}
\value{
A data frame. When compile is TRUE, ready to be imported into Epigraf.
Otherwise with rows in the attributes of the data frame.
}
\description{
Map a data frame to the Relational Article Model
}
\examples{
library(eprgraf)

df <- tribble(
  ~case, ~title, ~text,
  1, "Case 01", "Happy New Year!",
  2, "Case 02", "Happy Easter!",
  3, "Case 03", "Happy Birthday!"
)

df |>
 df_to_ram(
   item = c("content" = "text"),
   article = c("id" = "case", "signature" = "case", "name" = "title"),
   compile = T
 )

}
