% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{.craft_add_rows}
\alias{.craft_add_rows}
\title{Add a row to the rows in the epi attribute}
\usage{
.craft_add_rows(ds, rows, skip = FALSE)
}
\arguments{
\item{ds}{A tibble}

\item{skip}{Whether to update the record or only use it as reference}

\item{row}{A crafted row}
}
\value{
Epigraf tibble
}
\description{
Add a row to the rows in the epi attribute
}
