% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{.craft_stop}
\alias{.craft_stop}
\title{Stop and say how to proceed}
\usage{
.craft_stop(msg)
}
\arguments{
\item{msg}{The stop message, e.g. "Please, craft an article first"}
}
\description{
Stop and say how to proceed
}
