% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.to_epitable}
\alias{.to_epitable}
\title{Add the epi_tbl class and make it remember its source}
\usage{
.to_epitable(data, source = NULL)
}
\arguments{
\item{data}{A tibble}

\item{source}{A named vector of source parameters, containing endpount, parameters and database name}
}
\description{
Add the epi_tbl class and make it remember its source
}
