% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode}
\alias{encode}
\title{Converts numbers using the radix vector}
\usage{
encode(number, base)
}
\arguments{
\item{number}{The number to convert}

\item{base}{The base}
}
\description{
APL-encode / APL-representation "T", pw 10/02
}
