% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_create_empty_items}
\alias{epi_create_empty_items}
\title{Create one empty item for each section}
\usage{
epi_create_empty_items(sections, itemtype)
}
\arguments{
\item{sections}{A data frame containing the sections}

\item{itemtype}{A string with the name for the itemtype}
}
\description{
Create one empty item for each section
}
