% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{epi_create_properties}
\alias{epi_create_properties}
\title{Function to create properties}
\usage{
epi_create_properties(propertytype, lemmata, names = NA, irifragments = NA)
}
\arguments{
\item{propertytype}{The property type (character)}

\item{lemmata}{A vector of lemmata (character).}

\item{names}{Optional. A vector of  names (character).}

\item{irifragments}{Optional. A vector of IRI fragments (character).}
}
\description{
Creates properties table.
}
