% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_create_sections}
\alias{epi_create_sections}
\title{Create sections}
\usage{
epi_create_sections(data, sectiontype, name = NA)
}
\arguments{
\item{data}{A data frame containing the columns articletype and norm_iri}

\item{sectiontype}{A string with the sectiontype}

\item{name}{The section name. Leave empty to use the name defined in the Epigraf domain model.}
}
\description{
Creates a section for each row in the input data frame
}
