% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_extract_wide}
\alias{epi_extract_wide}
\title{Select nested data from prefixed columns}
\usage{
epi_extract_wide(data, cols_prefix, cols_keep = c())
}
\arguments{
\item{data}{A data frame}

\item{cols_prefix}{All columns with the prefix will be selected,
the prefix will be removed from the column name.}

\item{cols_keep}{Convert the provided column names to underscored columns}
}
\value{
A dataframe containing all columns with the prefix without the prefix
}
\description{
Select nested data from prefixed columns
}
