% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_is_id}
\alias{epi_is_id}
\title{Check whether the provided vector contains valid IDs prefixed
with table names. Example: articles-123}
\usage{
epi_is_id(ids, table = NA)
}
\arguments{
\item{ids}{The vector that will be proofed}

\item{table}{Check whether the path contains the table. Leave empty to allow all tables.}
}
\description{
Check whether the provided vector contains valid IDs prefixed
with table names. Example: articles-123
}
