% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_is_irifragment}
\alias{epi_is_irifragment}
\title{Check whether the provided vector contains a valid IRI fragment}
\usage{
epi_is_irifragment(irifragment)
}
\arguments{
\item{iripath}{The vector that will be proofed}
}
\description{
Check whether the provided vector contains a valid IRI fragment
}
