% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_is_iripath}
\alias{epi_is_iripath}
\title{Check whether the provided vector contains a valid IRI path}
\usage{
epi_is_iripath(iripath, table = NA, type = NA)
}
\arguments{
\item{iripath}{The vector that will be proofed}

\item{table}{Check whether the path contains the table. Leave empty to allow all tables.}

\item{type}{Check whether the path contains the type.  Leave empty to allow all types.}
}
\description{
Check whether the provided vector contains a valid IRI path
}
