% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_is_prefixid}
\alias{epi_is_prefixid}
\title{Check whether the provided vector contains valid IDs prefixed
with table names and temporary prefixes. Example: articles-tmp123}
\usage{
epi_is_prefixid(ids, table = NA, prefix = NA)
}
\arguments{
\item{ids}{The vector that will be proofed}

\item{table}{Check whether the path contains the table. Leave empty to allow all tables.}

\item{prefix}{Check whether the ID contains the prefix, e.g. "tmp". Leave empty to allow all prefixes.}
}
\description{
Check whether the provided vector contains valid IDs prefixed
with table names and temporary prefixes. Example: articles-tmp123
}
