% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_text2article}
\alias{epi_text2article}
\title{Convert text to epigraf article}
\usage{
epi_text2article(text)
}
\arguments{
\item{text}{Dataframe with the columns id, project, caption and content}
}
\value{
Dataframe with article, section and item
}
\description{
Convert text to epigraf article
}
\examples{
epi_text2article(
tibble(id=1,project="Import",
caption="My first text",
content="Words are letters with glue"))
}
