% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{epi_wide_to_long}
\alias{epi_wide_to_long}
\title{Convert wide to long format}
\usage{
epi_wide_to_long(data)
}
\arguments{
\item{data}{A dataframe with the column id containing a valid IRI path.
Additional columns may contain nested data in the following form:

\if{html}{\out{<div class="sourceCode">}}\preformatted{        Column names prefixed with "properties", "items", "sections",
         "articles" and "projects" followed by a dot (e.g. "properties.id",
        "properties.lemma") will be extracted and stacked to the dataframe.
}\if{html}{\out{</div>}}}
}
\value{
A dataframe with all input rows and the nested records stacked.
}
\description{
Convert wide to long format
}
