% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigraf-package.R
\docType{package}
\name{epigraf-package}
\alias{epigraf}
\alias{epigraf-package}
\title{epigraf: Convert text to epigraf article}
\description{
Convert canonical SM datasets to Epigraf structure for import. To learn more about epigraf, start with the vignettes: https://github.com/strohne/datavana/tree/main/epigraf.
}
\author{
\strong{Maintainer}: Jakob Jünger \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
