% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_items_to_ram}
\alias{map_items_to_ram}
\title{Create RAM rows for item data}
\usage{
map_items_to_ram(df, mapping)
}
\arguments{
\item{df}{The source data frame}

\item{mapping}{The mapping of source columns to RAM columns}
}
\value{
RAM rows
}
\description{
Create RAM rows for item data
}
\keyword{internal}
