% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_lists}
\alias{merge_lists}
\title{Merge list elements by their name}
\usage{
merge_lists(l)
}
\arguments{
\item{l}{A list of lists to merge}
}
\value{
A merged list
}
\description{
Merge list elements by their name
}
