% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na.omit.list}
\alias{na.omit.list}
\title{Remove NA entries from a list}
\usage{
\method{na.omit}{list}(y)
}
\arguments{
\item{y}{A list}
}
\value{
A cleaned list
}
\description{
Remove NA entries from a list
}
