% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{net_getnodes}
\alias{net_getnodes}
\title{Get all distinct nodes in an edge list}
\usage{
net_getnodes(edges, col_source, col_target)
}
\arguments{
\item{edges}{An edge list}

\item{col_source}{Source column name}

\item{col_target}{Target column name}
}
\value{
A tibble with one column \code{id} containing unique source and target values
}
\description{
Get all distinct nodes in an edge list
}
