% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num2abc}
\alias{num2abc}
\title{Convert a number to letters, e.g. 3 becomes c}
\usage{
num2abc(number, base = 26)
}
\arguments{
\item{number}{The number to convert}

\item{base}{The number of letters to use}
}
\description{
Convert a number to letters, e.g. 3 becomes c
}
