% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_json}
\alias{parse_json}
\title{Parse JSON columns}
\usage{
parse_json(data)
}
\arguments{
\item{data}{A character vector that may contain NAs}
}
\value{
A parsed vector
}
\description{
Parse JSON columns
}
