% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudonyms}
\alias{pseudonyms}
\title{Create distinct pseudonyms}
\usage{
pseudonyms(n)
}
\arguments{
\item{n}{Number of pseudonyms. Leave empty when used inside mutate to create value for each row}
}
\value{
Vector of pseudonyms
}
\description{
Create distinct pseudonyms
}
