% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram.R
\name{ram_compile}
\alias{ram_compile}
\title{Compile a crafted Epigraf table ready to patch into the database}
\usage{
ram_compile(ds)
}
\description{
Compile a crafted Epigraf table ready to patch into the database
}
