% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like2epi.R
\name{sm_canonical2epi}
\alias{sm_canonical2epi}
\title{Convert a social media dataset from its canonical form
to the form expected by Epigraf}
\usage{
sm_canonical2epi(data)
}
\arguments{
\item{data}{Datafram containing a social media dataset in its canonical form}
}
\value{
Dataframe containing project, articles, sections, items and properties rows
}
\description{
Convert a social media dataset from its canonical form
to the form expected by Epigraf
}
