% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_cleanhtml}
\alias{sm_cleanhtml}
\title{Fix xml attributes and entities}
\usage{
sm_cleanhtml(data, column)
}
\arguments{
\item{data}{data}

\item{column}{column}
}
\description{
Fix xml attributes and entities
}
