% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_create_articles}
\alias{sm_create_articles}
\title{Creates an article row from the first message in each thread}
\usage{
sm_create_articles(msg)
}
\arguments{
\item{msg}{Dataframe with the following columns:
platform, tree_thread, tree_pos,
caption, text,
sample_name, sample_no}
}
\description{
Creates an article row from the first message in each thread
}
