% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2epi.R
\name{sm_create_items}
\alias{sm_create_items}
\title{Create sections_id, articles_id, type and sortno fields}
\usage{
sm_create_items(msg, itemtype, sortno)
}
\arguments{
\item{msg}{message}
}
\description{
Create sections_id, articles_id, type and sortno fields
}
