% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2epi.R
\name{sm_create_metrics_items}
\alias{sm_create_metrics_items}
\title{Create items containing metrics}
\usage{
sm_create_metrics_items(msg, sortno)
}
\arguments{
\item{msg}{Dataframe}
}
\description{
Create items containing metrics
}
