% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_create_projects}
\alias{sm_create_projects}
\title{Create project rows}
\usage{
sm_create_projects(msg)
}
\arguments{
\item{msg}{Dataframe with the columns:
sample_name, ...}
}
\description{
Creates a project row for each sample
}
