% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like2epi.R
\name{sm_create_properties}
\alias{sm_create_properties}
\title{Create properties from a column}
\usage{
sm_create_properties(msg, col_id, col_lemma)
}
\arguments{
\item{msg}{Dataframe}
}
\description{
Create properties from a column
}
