% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like2epi.R
\name{sm_create_property_items}
\alias{sm_create_property_items}
\title{Create items linked to a property}
\usage{
sm_create_property_items(msg, col, sortno, first = T)
}
\arguments{
\item{msg}{Dataframe}
}
\description{
Create items linked to a property
}
