% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2epi.R
\name{sm_create_search_items}
\alias{sm_create_search_items}
\title{Create items from caption and text for the full text search}
\usage{
sm_create_search_items(msg, sortno)
}
\arguments{
\item{msg}{Dataframe}
}
\description{
Create items from caption and text for the full text search
}
