% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2epi.R
\name{sm_create_sections}
\alias{sm_create_sections}
\title{Creates a section for each message in a thread}
\usage{
sm_create_sections(msg)
}
\arguments{
\item{msg}{Dataframe with the following columns:
platform, tree_thread, tree_pos,
caption, text,
sample_name, sample_no}
}
\description{
Creates a section for each message in a thread
}
