% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_create_text_items}
\alias{sm_create_text_items}
\title{Create items with the message content}
\usage{
sm_create_text_items(msg, sortno)
}
\arguments{
\item{msg}{Dataframe}
}
\description{
Create items with the message content
}
