% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_pseudonyms}
\alias{sm_pseudonyms}
\title{Anonymize authors}
\usage{
sm_pseudonyms(msg)
}
\arguments{
\item{msg}{Dataframe with the columns:
platform, author_id, ...}
}
\value{
Dataframe with the columns platform, author_id, author_name, ...
}
\description{
Anonymizing authornames
}
