% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_unnest_property_items}
\alias{sm_unnest_property_items}
\title{Create property items from a column with semicolon separated values}
\usage{
sm_unnest_property_items(msg, col_id, sortno, first = F)
}
\arguments{
\item{msg}{Dataframe}
}
\description{
Create property items from a column with semicolon separated values
}
