% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2epi.R
\name{sm_unnest_property_properties}
\alias{sm_unnest_property_properties}
\title{Create properties from a column with semicolon separated values}
\usage{
sm_unnest_property_properties(msg, col_id)
}
\arguments{
\item{msg}{Dataframe}
}
\description{
Create properties from a column with semicolon separated values
}
