% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_text2article.R
\name{textToArticle}
\alias{textToArticle}
\title{Convert text to epigraf article}
\usage{
textToArticle(text)
}
\arguments{
\item{text}{Dataframe with the columns id, project, caption and content}
}
\value{
Dataframe with article, section and item
}
\description{
Convert text to epigraf article
}
\examples{
textToArticle(
tibble(id=1,project="Import",
caption="Mein erster Text",
content="Worte sind Buchstaben mit Kleber"))
}
