% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_add_level2}
\alias{tree_add_level2}
\title{Add level, thread and order}
\usage{
tree_add_level2(data, col_id, col_parent, col_sort = NULL)
}
\arguments{
\item{col_id}{The ID column of the node}

\item{col_parent}{The ID column of the parent node}

\item{col_sort}{Column for sorting the nodes inside each parent. Leave empty to use the ID column.}

\item{.data}{The dataframe containing hierarchical data}
}
\value{
Data frame with the additional columns tree_thread, tree_order and tree_level
}
\description{
TODO: fix tree_order ... should be related to the thread, not the parent
}
\details{
Version 2: loop through the levels
}
