% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_add_mptt}
\alias{tree_add_mptt}
\title{Adds left and right values to the dataframe}
\usage{
tree_add_mptt(.data)
}
\arguments{
\item{data}{Dataframe with the columns tree_id, tree_parent, tree_thread, tree_level, tree_order}
}
\value{
Dataframe with lft and rght values
}
\description{
Left and right values are used to store hierarchical data
using the concept of modified preorder tree traversal.
See https://www.sitepoint.com/hierarchical-data-database-3/
}
