% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_add_path}
\alias{tree_add_path}
\title{Add a column holding the path of each node.}
\usage{
tree_add_path(data, col_id, col_parent_id, col_lemma, delim = "/")
}
\arguments{
\item{data}{Dataframe containing hierarchical data}

\item{col_id}{The ID column of the node}

\item{col_lemma}{The column holding the node name that will be used for the path}

\item{delim}{Character that glues together the path elements. Set to NULL to create a vector instead.}

\item{col_parent}{The ID column of the parent node}
}
\value{
A data frame with the additional column tree_path
}
\description{
The path is created by concatenating all col_lemma values up to the root node.
Lemmata are concatenated using a slash - existing slashes are replaced by the entity /.
}
