% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{tree_bind_ancestors}
\alias{tree_bind_ancestors}
\title{Row bind all ancestors of the selected nodes}
\usage{
tree_bind_ancestors(.data, .tree, id, parent_id)
}
\arguments{
\item{.data}{Data frame containing the selected nodes}

\item{.tree}{Data frame containing all nodes including all ancestors}

\item{id}{Column name of the id in .data and .tree}

\item{parent_id}{Column name of the parent id in .data and .tree}
}
\value{
Data frame containing the nodes of .data and all ancestors
}
\description{
Row bind all ancestors of the selected nodes
}
