% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch}
\alias{api_patch}
\title{Patch data}
\usage{
api_patch(data, database, table = NA, type = NA, wide = T)
}
\arguments{
\item{data}{A dataframe with the column \code{id}.
Additional columns such as norm_data will be written to the record.
The id must either be a a valid IRI path (e.g. properties/objecttypes/xxx)
or an id prefixed by the table name (e.g. properties-12).
Patching properties with prefixed ids requires a \code{type} column
that contains the property type.
Column names with table names as prefixes will be extracted, if wide is set to TRUE (default).}

\item{database}{The database name}

\item{table}{Optional: Check that the data only contains rows for a specific table}

\item{type}{Optional: Check that the data only contains rows with a specific type}

\item{wide}{Convert wide format to long format.
If TRUE, column names prefixed with "properties", "items", "sections", "articles"
and "projects" followed by a dot (e.g. "properties.id",
"properties.lemma") will be extracted and patched as additional records.}
}
\description{
Update records in the database using the API.
Existing records will be updated, missing records will be created.
The function supports uploading all data related to articles:
articles, sections, items, links, footnotes, properties, projects, users, types.
The IRI path in the ID column of the dataframe must contain the specific table name.
}
