% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_patch_items}
\alias{api_patch_items}
\title{Patch items}
\usage{
api_patch_items(items, database)
}
\arguments{
\item{items}{Dataframe with the columns
article (must be a valid IRI path),
section (either a valid IRI path or a section type),
item (the item type).
sortno (to address multiple items in a section, optional, defaults to 1)
Additional columns such as properties_id or value will be
written to the item.}

\item{database}{The database name}
}
\description{
Existing items will be updated, missing items will be created.
The section IRI fragments and the item IRI iri fragments will be derived from the
article IRI fragments. The item IRI will be suffixed by the sortno value (default=1).
}
