% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_setup}
\alias{api_setup}
\title{Save API connection settings to environment variables.}
\usage{
api_setup(apiserver, apitoken, verbose = F)
}
\arguments{
\item{apiserver}{URL of the Epigraf server (including https-protocol)}

\item{apitoken}{Access token}

\item{verbose}{Show debug messages and the built URLs}
}
\description{
Save API connection settings to environment variables.
}
