% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_table}
\alias{api_table}
\title{Download tables}
\usage{
api_table(endpoint, params = c(), db, maxpages = 1, silent = FALSE)
}
\arguments{
\item{endpoint}{The endpoint path (e.g. "articles/index" or "articles/view/1")}

\item{params}{A named list of query params}

\item{db}{The database name}

\item{maxpages}{Maximum number of pages to request.
Set to 1 for non-paginated tables.}

\item{silent}{Whether to output status messages}
}
\description{
Fetches tables such as articles, projects or properties
}
\details{
TODO: silent problem message (false positive, results from the last column being empty from case to case)
TODO: add progress bar
}
