% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft_article}
\alias{craft_article}
\title{Craft an article}
\usage{
craft_article(
  ds,
  col_id,
  col_name = "",
  col_signature = "",
  col_sortno = "",
  type = "default"
)
}
\description{
Craft an article
}
