% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft_item}
\alias{craft_item}
\title{Craft an item}
\usage{
craft_item(
  ds,
  col_id,
  col_content = "",
  type_property = "default",
  col_property = "",
  col_sortno = "",
  type = "default"
)
}
\description{
Craft an item
}
