% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft_section}
\alias{craft_section}
\title{Craft a section}
\usage{
craft_section(
  ds,
  col_id,
  col_name = "",
  col_alias = "",
  col_sortno = "",
  type = "default"
)
}
\description{
Craft a section
}
