% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_table}
\alias{db_table}
\title{Get tables from a database}
\usage{
db_table(table, db, deleted = FALSE)
}
\arguments{
\item{table}{Table name}

\item{db}{A connection object (object) or the database name (character)}
}
\description{
Get tables from a database
}
