% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigraf-package.R
\docType{package}
\name{epigraf-package}
\alias{epigraf}
\alias{epigraf-package}
\title{epigraf: Convert text to epigraf article}
\description{
Convert canonical social media datasets to Epigraf structure for import. To learn more about epigraf, start with the vignettes on Github.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/strohne/datavana/tree/main/epigraf}
}

}
\author{
\strong{Maintainer}: Jakob Jünger \email{jakob.juenger@uni-muenster.de} (\href{https://orcid.org/0000-0003-1860-6695}{ORCID})

}
\keyword{internal}
