% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigraf-package.R
\docType{package}
\name{epigraf-package}
\alias{epigraf}
\alias{epigraf-package}
\title{epigraf: Fetch from and push data to Epigraf}
\description{
Provides functions to import and export data and for batch operations. To learn more about Epigraf, start with the vignettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/strohne/datavana/tree/main/epigraf}
}

}
\author{
\strong{Maintainer}: Jakob Jünger \email{jakob.juenger@uni-muenster.de} (\href{https://orcid.org/0000-0003-1860-6695}{ORCID})

}
\keyword{internal}
